Rename README.md to Elimuhub_README.md
This commit message describes the changes made to the README file of the project. The file name has been changed from README.md to Elimuhub_README.md to avoid confusion with other files in the same directory. The README file contains information about the project, such as:

- The goal and vision of Elimuhub, which is to provide a platform for online learning and collaboration in Kenya
- The main features of Elimuhub, such as courses, quizzes, forums, chat, and video conferencing
- The functionalities of Elimuhub, such as creating and enrolling in courses, taking quizzes, posting and commenting on forums, chatting with other users, and joining video calls
- The technologies and tools used to develop Elimuhub, such as HTML, CSS, JavaScript, Bootstrap, Node.js, Express, MongoDB, Socket.io, and WebRTC

The commit message follows the best practices of writing clear and concise messages that explain the rationale behind the changes.